updating snapshots
