+++ projects/standard-rulesets/src/breaking-changes/helpers/rangeChange.ts
@@ -49,52 +49,3 @@ export function numericMinimumRangeShouldNotExpand(
     }
   }
 }
-
-export function isArrayProperty(
-  before: OpenApi3SchemaFact,
-  after: OpenApi3SchemaFact
-) {
-  const numericTypes = ['array'];
-  if (typeof before.type === 'string' && typeof after.type === 'string') {
-    return (
-      numericTypes.includes(before.type) && numericTypes.includes(before.type)
-    );
-  }
-  return false;
-}
-
-export function arrayMaxItemsShouldNotIncrease(
-  before: OpenApi3SchemaFact,
-  after: OpenApi3SchemaFact
-) {
-  if (isArrayProperty(before, after)) {
-    if (
-      typeof before.maxItems !== 'undefined' &&
-      typeof after.maxItems !== 'undefined'
-    ) {
-      if (before.maxItems < after.maxItems) {
-        throw new RuleError({
-          message: `array maxItems  increased: ${before.maxItems} -> ${after.maxItems}`,
-        });
-      }
-    }
-  }
-}
-
-export function arrayMinItemsShouldNotDecrease(
-  before: OpenApi3SchemaFact,
-  after: OpenApi3SchemaFact
-) {
-  if (isArrayProperty(before, after)) {
-    if (
-      typeof before.minItems !== 'undefined' &&
-      typeof after.minItems !== 'undefined'
-    ) {
-      if (before.minItems > after.minItems) {
-        throw new RuleError({
-          message: `array minItems decreased: ${before.minItems} -> ${after.minItems}`,
-        });
-      }
-    }
-  }
-}
